Order Routing IOL
Former-commit-id: 17683380f020ee53ae868adad0e647e86445b9e3

+++ HFTPlatform/zHFT.OrderRouters.InvertirOnline.Common/DTO/BaseOrder.cs
@@ -14,7 +14,7 @@ namespace zHFT.OrderRouters.InvertirOnline.Common.DTO
 
         public string simbolo { get; set; }
 
-        public long cantidad { get; set; }
+        public double cantidad { get; set; }
 
         public double? precio { get; set; }
